Disposition Effect

Compute the disposition effect in a vectorized way.

disposition_effect(realized_gains, paper_gains, realized_losses, paper_losses)

disposition_difference(gains, losses)

disposition_compute(gainslosses, aggregate_fun = NULL, ...)

disposition_compute_ts(gainslosses, aggregate_fun = NULL, ...)

disposition_summary(gainslosses)

disposition_summary_ts(de_timeseries)

Arguments

realized_gains
Numeric vector (or scalar) containing realized gains values.

paper_gains
Numeric vector (or scalar) containing paper gains values.

realized_losses
Numeric vector (or scalar) containing realized losses values.

paper_losses
Numeric vector (or scalar) containing paper losses values.

gains
Numeric vector (or scalar) containing gains.

losses
Numeric vector (or scalar) containing losses.

gainslosses
Data frame, the portfolio of the investor containing the realized and paper gains and losses results (as those obtained via portfolio_compute).

aggregate_fun
Function to use to aggregate results. Default to NULL, that is no aggregation is performed and the results of each asset are shown.

...
Further arguments to be passed to the aggregate function.

de_timeseries
Data frame, the time series of disposition effects.

Value

Numeric vector (or scalar) with the value(s) of disposition effect(s) or disposition difference(s). A tibble is returned by ................

Details

The disposition effect is defined as \(DE = (Realized Gain / (Realized Gain - Paper Gain)) - (Realized Loss / (Realized Loss + Paper Loss))\)

The disposition difference is defined as \(DD = Realized Gain - |Realized Loss|\) or \(DD = Paper Gain - |Paper Loss|\)

Functions

disposition_effect: Compute the disposition effect as defined by L. Mazzucchelli et al.

disposition_difference: Compute the disposition difference as defined by L. Mazzucchelli et al.

disposition_compute: Compute the disposition effect directly on the investor's portfolio containing realized and paper gains and losses results.

disposition_compute_ts: Compute the time series disposition effect on the gains and losses results.

disposition_summary: Wrapper that returns the most important summary statistics related to the disposition effect.

disposition_summary_ts: Wrapper that returns the most important summary statistics related to the time series disposition effect.

References

H. Shefrin & M. Statman, 1985

Author

L. Mazzucchelli & M. Zanotti

Examples

  data(portfolio_results)

  # Disposition effect on count for each traded asset
  de <- disposition_effect(portfolio_results$RG_count,
                           portfolio_results$PG_count,
                           portfolio_results$RL_count,
                           portfolio_results$PL_count)
  names(de) <- portfolio_results$asset

  # Average disposition effect of the investor
  mean(de)

#> [1] -0.3047619

  # Realized difference on value for each asset
  disposition_difference(portfolio_results$RG_value, portfolio_results$RL_value)

#> [1]  0.008701958 -0.074978204 -0.030965392  0.010551792  0.000000000

  # Paper difference on value for each asset
  disposition_difference(portfolio_results$PG_value, portfolio_results$PL_value)

#> [1] 0.03002533 0.00000000 0.00000000 0.13492566 0.00000000